refactor(daemon): extract SemanticNeighborGraphBuilder from EmbeddingService
updateSemanticNeighborGraphUnlocked() was a 745-line method on
EmbeddingService that owned its own caches, counters, and three
environment knobs, and could only be exercised through a live embed
job. It moves verbatim into SemanticNeighborGraphBuilder
(src/daemon/components/semantic_neighbor_graph_builder.cpp), which
talks to the outside through the KG store, the vector database, an
EdgeSink for persisting edges, and an optional phase timer. Its caches
and the edges/docs/errors counters move with it; EmbeddingService
keeps the mutation lock, the pending-by-model bookkeeping, and the
idle backfill, and supplies the WriteCoordinator-backed sink and the
timing sink from its constructor.

The three YAMS_GRAPH_SEMANTIC_* environment reads become typed config:
[embeddings.semantic_graph] top_k, similarity_threshold, use_hnsw,
resolved by ConfigResolver::applyEmbeddingSemanticGraph into
DaemonConfig::embeddingService.semanticGraph (the daemon refuses to
start on an unknown key or an out-of-range value, like the other typed
sections). TopologyManager folded the two env values into its cell-identity
string; it now takes the typed policy through its Dependencies. That
string is telemetry only (lastCellIdentity and the rebuild result), so
its change from "topk=default" to the explicit value has no operational
effect. The parse boundary is stricter than the env reads were: top_k
= 0, a non-integer, a threshold outside [0, 1], or a non-boolean
use_hnsw now stops the daemon at startup instead of being clamped or
ignored. The retrieval-quality bench drops the retired names as fallback
spellings for its own YAMS_BENCH_SEED_* seeds; production_environment_keys.txt
loses the three keys.

Tests: semantic_neighbor_graph_builder_catch2_test.cpp drives the
builder against an in-memory vector database and a sqlite KG store
with the edge sink captured: the streaming and corpus-wide passes
emit the same 16 forward/reverse edges with cosine weights above the
threshold, top-K and the adaptive floor are honoured, and a missing
writer is counted (twice, because the dropped streaming write falls
through to the corpus pass; pre-existing and pinned, not changed).
config_resolver_test.cpp covers defaults, explicit values, and
rejection. The daemon component and topology manager suites pass
unchanged; the policy suite is green with the allowlists updated.

EmbeddingService.cpp: 3,161 -> 2,419 lines.
