+++ src/daemon/meson.build
@@ -126,6 +126,7 @@ yams_daemon_lib = static_library('yams_daemon',
     'components/embed_preparer.cpp',
     'components/IOCoordinator.cpp',
     'components/EmbeddingService.cpp',
+    'components/semantic_neighbor_graph_builder.cpp',
     'components/TuningManager.cpp',
     'components/GradientLimiter.cpp',
     'components/ResourceGovernor.cpp',
