+++ tests/scripts/production_environment_keys.txt
@@ -128,9 +128,6 @@ YAMS_GRADIENT_MIN_LIMIT
 YAMS_GRADIENT_SMOOTHING_ALPHA
 YAMS_GRADIENT_TOLERANCE
 YAMS_GRADIENT_WARMUP_SAMPLES
-YAMS_GRAPH_SEMANTIC_THRESHOLD
-YAMS_GRAPH_SEMANTIC_TOPK
-YAMS_GRAPH_SEMANTIC_USE_HNSW
 YAMS_GREP_MAX_DOCS_COLD
 YAMS_GREP_MAX_DOCS_HOT
 YAMS_GREP_MAX_RESULTS
